CourseAdmin add new course tweaks.
Use the course_admin_id instead of 'admin' when talking about the
admin course name on this page and in its help.

Mention setting $permissionLevels{login}='admin' in course.conf
in the help as a way to control access to the admin course.

Use `overflow-auto` and `max-height` in the select which admin users
to add to the new course div.