Allow adding multiple initial users when adding a new course.
When adding a new course, there is now an 'Add Another Instructor' button
which can be used to add more than one new user to the newly created course.
Each time this button is hit, the form expands allowing input for a new user.
It is possible to set the permission level of each of these users and to keep
the form balanced, the student ID can now also be included.

This also updates assigning the initial user to sets and achievements. The
intention was to assign new instructors to these, so any user (either copied
from the admin course or added initially) whose permission is less than admin
will be assigned to the sets and achievements.