Hotfix: Prevent error when adding new course.
When adding a new course if the initial instructor entered into
the add course form has the same user ID as a user in the course
non-student users, sets, and achievements are being copied from,
this user gets assigned to sets and achievements twice, causing
an error.

This skips assigning the initial user to the original sets they
may have been assigned to and will always assign the initial user
to all sets and achievements, so no error occurs.