Rework of LTI grade pass back.
Restructure LTI grade pass back so that there are no additional database
queries with the new restrictions on when grades are passed back to the
LMS.  With this restructuring the only database queries are made while
grading the sets the same as before.

The grade_set and grade_gateway methods return the things they pull from
the database so that the database records can be culled for the
additional information needed to determine if grade pass back should
occur.  The usage of these methods elsewhere is the same.  The new data
that is passed back will simply be ignored in those cases.

The `lib/WeBWorK/Authen/LTI/MassUpdate.pm` file has been renamed to
`lib/WeBWorK/Authen/LTI/GradePassback.pm` and contains the previous
`mass_update` method as well as all methods needed for the new
processing.

The `grade_all_sets` method that was in `lib/WeBWorK/Utils/Sets.pm` now
takes `$getSetGradeConditionally` as an optional last argument.  That
argument should be a reference to a subroutine that takes the arguments
$db, $ce, $studentName, and $userSet, and either returns a reference to
a hash containing the keys totalRight and total with the grade for the
set, or returns `undef`. If it returns `undef` then the set will not be
included in the grade computation.  Otherwise the totalRight and total
will be added into the grade for all sets.  If the optional last
arugment is not provided, then a default method will be used that does
the same thing as before.

Note that the `$LTI{v1p1}{grader}` option the `authen_LTI_1_1.conf.dist`
and `authen_LTI_1_3.conf.dist` files has been removed.  Instead the
correct module is determined directly from the `$LTIVersion` as was
previously done in the `lib/Mojolicious/WeBWorK/Tasks/LTIMassUpdate.pm`
file.  This way a system adiminstrator cannot accidentally delete that
or modify it in some way and mess things up.  Since this is only done in
these two places and not needed elsewhere the convenience of the course
environment variable containing the module name is not needed.