Eliminate the visually hidden spans injected by javascript.
The spans are simply added where needed.  No javascript is needed for
them to be in the page.

The "?" help label is changed to just "Help" in most cases.  The "?" is
usually not read by screen readers.  Note that these all go through the
`helpMacro` method and a help file template is rendered.  The template
can also set the ariaLabel stash value to override the default "Help"
text.

Most pages shouldn't use the override though (I think).  For example, on
the "Accounts Manager" page the screen reader says something like
"accounts manager heading level 1 link help link" (at least that is what
Gnome Orca says for this). I don't think it should be read as "account
manager heading level 1 link accounts manager help link".  It increases
the verbosity unnecessarily.

However, on the "Assignments" page the title of the page is the course
title.  So the default aria label is overridden with "Assignments Help",
and the screen reader says something like "course title heading level 1
link assignments help link".  Note the help title in the dialog was also
changed to "Assigments Help".  It still said "Course Home Help". It
seems we missed this when the page name was changed.

Of course mileage will vary as to what different screen readers say on
different systems and with different browsers.