Switch to using Perl::Tidy version 20240903 (the latest).
This commit just makes the changes in the scripts, but does not yet run
perltidy with the new setup.  That will be done in the next commit.

Note that the `-vxl='q'` option has been removed.  This option disables
vertical alignment of qw quotes.  Alignment of qw quotes was added in
version 20220613 (the previous version), and this option to disable that
was added to the webwork2 (and pg) perltidy configuration to avoid the
alignment because it changes so many files. However, I don't think that
alignment is bad, and I am tired of adding new options to prevent code
changes.  With this newer version another option would be needed to
prevent the new signed number alignment (although that affects PG much
more than webwork2).

Note that the `bin/dev_scripts/run-perltidy.pl` script now insists on
this specific version of perltidy instead of this version or newer.
This is because every time a new version of perltidy comes out, it
changes things.  So developers really need to be using the same version
or it will not be consistent with the github workflow result.

Also, the version requirement in the check_modules.pl script has been
removed.  Only developers need a specific version of perltidy.  Those
running webwork2 can use other versions.  This is only used for
formatting problems in the PG problem editor, and using other versions
may result in slightly different formatting there, but that is not so
consequential that we need to require a specific version. Note that this
change depends on the corresponding pg pull request, because in that
pull request the `-vxl='q'` option has been removed.  Otherwise version
20220613 or newer of perltidy would still be needed.