Add hidden span labels for help links.
This adds a hidden span that gives a label for the help links for
the main help link using the helpMacro and the help links on the
course configuration page.

The page title is used for the description for the main help link.
For the course configuration links, the setting short description
is used for the title. A few of the short descriptions ended in a
period, those were all removed to make the description "desc help."
consistent and not have a period before the help statement.