Don't limit macros to predefined lists.
To keep from having to update the list each time a macro directory
is added/removed/changed, just sort the macros by their directory
alphabetically. There is still a hash of macro names to allow both
translated or custom names instead of using the directory as its
name, but any directory that isn't in the list will also be included
without updating the list.