Rework confirmation to creating/recording answers when acting.
When acting as another user with permissions to create or record
answers as that user, rework how the confirmation message is
stored. Instead of piggy backing off of the `invalidSet` key,
use a new `actingConformation` key for this case.

This also skips deleting any proctor authorization key while
waiting for user confirmation to continue.