Make JITAR disabled problems disabled without javascript.
There is no reason to use javascript for this.  Just make the links
disabled to begin with.  Furthermore, there is no reason to use jQuery
for this.  I thought all webwork2 jQuery usage had already been removed,
but this was missed.

To test this create a JITAR set with two or more problems (or really two
or more nested sets of problems), and set "Restrict Problem Progression"
to "Yes".  Assign the set to a student user, and then log in as that
student user.  On the problem list page the latter prolems in the set
should be grayed out, and clicking on them does nothing (that already
works without javascript because these aren't even `a` tags, they are
`span` tags).  If you open the first problem, then the latter problems
in the prroblems list in the site menu should be grayed out, and
clicking on them does nothing. Currently, that uses jQuery javascript to
disable those links.  With this pull request, they are just disabled to
begin with.