Move achievement item use form to ProblemSet page.
Move the forms to use achievement items to a modal on the problem
set page.  If any achievement item is usable on the current set,
a button to "Use Achievement Reward" will be available, and open
a modal which lists all achievement items that can be applied.

The Achievements page still lists all rewards the student has earned,
and provides instructions to use them from the problem assignment they
wish to use the achievement on.