Updates to achievement items from feedback.
Remove the fs-3 class from the achievement item names on the
achievements page when listing the achievement items.

When printing the achievement item name with the number of remaining
items, don't treat separately from any > 0 value.

Update the problem selection drop down forms to be clearer on
what the problem number is for and clarifying when double means
double the weight vs increase grade by 50%.