Actually handle failure to find a problem from a grouping set gracefully.
Actually, this handles any failure to generate a new test version
gracefully. This wraps the set version assignment call in an `eval` and
catches any exceptions that occur in the process including issues from
an instructor not setting up grouping sets correctly.  If any exceptions
do occur the user is shown the usual invalid set screen with a message
stating, "Unable to generate a valid test version. This is usually
caused by invalid usage of grouping sets or a database error. Please
speak to your instructor to fix the error. A system administrator can
obtain more details on this error from the logs." The actual exception
is logged to the webwork2 application log file (logs/webwork2.log) so an
instructor can ask a system administrator for details of the error. Note
this does not add to the general webwork2 issue of showing exceptions in
the user interface.

This is one way to resolve issue #2669 and probably the best way without
massively revising the GatewayQuiz module.