Switch the feedback emails to HTML.
The email content is generated by the `templates/ContentGenerator/Feedback/feedback_email.html.ep`
 template. All of the same content is there, just arranged nicely into
tables and such. The link to the page the user sent the email from is of
course an `a href`.  For now it reads "To visit the page from which
(user name) sent feedback, click here." where "click here" is the link.
Note that it used to read "To visit the page from which the user sent
feedback, long link."  We have the user name so why not use it here?

Also, when the sent message is displayed for the user in the browser,
the message is formated much the same as it now is in the email.  That
is it is not wrapped and placed in a `pre` tag.  Instead it is in a
bordered `div` formatted with essentially the same style as in the
email.