PG Problem Editor file selector
This changes the PG Problem Editor to not assume a file type when opened
without a file type specified.  So when the problem editor is opened
from the site navigation link it does not directly open to the blank
file template.  Instead it opens to a file chooser page. On that page,
three options are presented.  The user may choose a "New Problem
Template", a "File" and give a path relative to the course templates
directory, or a "Sample Problem" and pick a sample problem file from a
dropdown.

In order to support sample problems a new file type needed to be
implemented for the problem editor.  The sample problem parsing module
for PG is used to strip out the documentation when the problem is loaded
into the editor.

However, I observed that the current code for that used by the sample
problem viewer to offer a sample problem file for download is highly
innefficient.  It currently reads all of the sample problem files, then
reads all of the PG macros, and then parses the desired problem
stripping out documentation, and finally serves the file. So a
corresponding PG pull request adds a `getSampleProblemCode` method that
simply reads only the desired sample problem and efficiently strips out
the documentation without needing the macros and sample problem
metadata. So instead of needing to read 253 files (the number of macros
plus the number of sample problems currently), it only reads 1.