Update the pg-codemirror-editor to pull in a bugfix.
There is early termination code in the PGML parser that is not actually
part of PGML. It is there to catch impartial constructs that occur while
typeing.  For instance, say you have already typed ``[` `]``, and then
you start to type `[$` inside that.  This code terminates the outer
block so that syntax highlighting continues to work after the outer
block.

Unfortunately, it was also early terminating 'balance' blocks inside
outer blocks even though those could contain the terminator of the outer
block.  This can cause tokens inside the block to be incorrectly
terminated early.  For example, in `[@ '[@ "inner code" @]' @]**` the
single quote starts a 'balance' type block, and the `@]` inside the
string tries to terminate the outer `[@` block. Of course that example
isn't something you should do any, but this was seen in a more
complicated construct than that example.  That sort of thing is often
needed for cases inside a BEGIN_PGML_SOLUTION/END_PGML_SOLUTION block.