Instead of a local `die` handler share a variable to the safe compartment.
The `local $SIG{__DIE__}` handler causes an issue with the
`Mojo::Reactor::EV` package and leaves the app in an unrecoverable
state.  So don't use that.  Instead share a localized package array
variable.  When `include` has errors it pushes a message onto that
array.  The first thing on that array is used for the exception (so the
first exception that occurs).