Rewrite the WeBWorK::Upload module.
This takes advantage of things available with Mojolicious.

This does not use `Data::UUID` anymore.  A comment stated that it was
overkill.  It is and it was never a good idea to use that for this
purpose.  This just uses the `Mojo::Asset::Memory::to_file` method which
converts to a `Mojo::Asset::File` object and saves the file to a
temporary file using `Mojo::File::tempfile` (which uses `File::Temp`
under the hood).  That is guaranteed to give a safe filename that does
not conflict with any existing files.  Setting the environment variable
`MOJO_TMPDIR` locally to be the passed in directory (which is
`$ce->{webworkDirs}{uploadCache}`) ensures that the file is saved in the
usual webwork2 upload location (although there is no need for that, and
we could drop that configuration variable and just use the system
temporary directory for this).

The info file that is written is now UTF-8 encoded and decoded.  This
fixes issue #2690.

The `dir` option which was previously required and the only "option" is
now just a required argument.  It doesn't make sense to use an "option"
hash things that are required.  Particularly if there is only one option
and it is required.

Other than what is mentioned above the module behaves much the same.