Fix random order page change issue.
I bet you thought all of the random order gateway test issues were
resolved.  Nope.

When a page change occurs in a random order test, the answers saved in
the hidden inputs for the problems not on the current page are still not
respecting the order.  As a result the answers get moved around to
incorrect problems each time that a page change occurs.